[chore] Fruits, Juice를 별도의 타입으로 분리
