+++ JuiceMaker/JuiceMaker/Model/Fruits.swift
@@ -1,0 +1,16 @@
+//
+//  Fruits.swift
+//  JuiceMaker
+//
+//  Created by J.E on 2023/01/03.
+//
+
+import Foundation
+
+enum Fruits: CaseIterable {
+    case strawberries
+    case bananas
+    case kiwis
+    case pineapples
+    case mangos
+}
